workers/vector/pyproject.toml
[project]
name = "timbermap-vector-worker"
version = "0.1.0"
description = ""
authors = [
    {name = "timbermap",email = "sebastian@timbermap.com"}
]
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "google-cloud-storage (>=3.10.0,<4.0.0)",
    "google-cloud-secret-manager (>=2.26.0,<3.0.0)",
    "google-cloud-pubsub (>=2.36.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "geopandas (>=1.1.3,<2.0.0)",
    "pyproj (>=3.7.2,<4.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "fiona (>=1.10.1,<2.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "geoalchemy2 (>=0.18.4,<0.19.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
